+++ tests/httpInterface.utils.js
@@ -35,7 +35,8 @@ async function setUpTest(
   testFixtures = fixtures,
   mongoDBCollectionName = BOOKS_COLLECTION_NAME,
   exposeMetrics = false,
-  strictOutputValidation = false
+  strictOutputValidation = false,
+  openApiSpecification = 'swagger'
 ) {
   const databaseName = getMongoDatabaseName()
   const mongoURL = getMongoURL(databaseName)
@@ -58,6 +59,7 @@ async function setUpTest(
     // Here is not in lower case for testing it
     USER_ID_HEADER_KEY: 'userId',
     VIEWS_DEFINITION_FOLDER: join(__dirname, 'viewsDefinitions'),
+    OPEN_API_SPECIFICATION: openApiSpecification,
   }
 
   const fastify = await lc39('./index', { logLevel, envVariables, exposeMetrics })